Moved backend to backend folder, added endpoint for getting available spaces
