+++ .gitignore
@@ -1,2 +1,4 @@
 .env
 env/
+backend/env/
+backend/__pycache__/
